voice_service/requirements.txt
fastapi
uvicorn
openai-whisper
soundfile
torch
